alon-client/pyproject.toml
[tool.poetry]
name = "alon-client"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-okx = "^0.3.5"
websockets = "^14.1"
ccxt = "^4.4.49"
pyyaml = "^6.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
